:recycle: refactor: rename directory - lv01 -> lv02 - lv02 -> lv03
