--- src/main/java/com/innovation/spring/entry/memo/lv02/dto/request/AuthSignUpRequest.java
@@ -1,13 +1,0 @@
-package com.innovation.spring.entry.memo.lv02.dto.request;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record AuthSignUpRequest(
-        @NotBlank @Email(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$")
-        String email,
-
-        @NotBlank
-        String password
-) {
-}
