--- src/main/java/com/innovation/spring/entry/memo/lv02/dto/response/AuthSignInResponse.java
@@ -1,9 +1,0 @@
-package com.innovation.spring.entry.memo.lv02.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record AuthSignInResponse(
-        String accessToken
-) {
-}
