--- src/main/java/com/innovation/spring/entry/memo/lv02/entity/User.java
@@ -1,31 +1,0 @@
-package com.innovation.spring.entry.memo.lv02.entity;
-
-import com.innovation.spring.entry.memo.global.entity.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "users")
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class User extends BaseTimeEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Builder
-    public User(String email, String password) {
-        this.email = email;
-        this.password = password;
-    }
-}
