--- src/main/java/com/innovation/spring/entry/memo/lv02/exception/ExceptionResponse.java
@@ -1,11 +1,0 @@
-package com.innovation.spring.entry.memo.lv02.exception;
-
-import lombok.Builder;
-import org.springframework.http.HttpStatus;
-
-@Builder
-public record ExceptionResponse(
-        HttpStatus status,
-        String message
-) {
-}
