--- src/main/java/com/innovation/spring/entry/memo/lv02/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.innovation.spring.entry.memo.lv02.repository;
-
-import com.innovation.spring.entry.memo.lv02.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    boolean existsByEmail(String email);
-
-    Optional<User> findByEmail(String email);
-
-}
