+++ src/main/java/com/innovation/spring/entry/memo/lv03/dto/request/AuthSignInRequest.java
@@ -1,0 +1,13 @@
+package com.innovation.spring.entry.memo.lv03.dto.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record AuthSignInRequest(
+        @NotBlank @Email(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$")
+        String email,
+
+        @NotBlank
+        String password
+) {
+}
