+++ src/main/java/com/innovation/spring/entry/memo/lv03/dto/response/AuthSignInResponse.java
@@ -1,0 +1,9 @@
+package com.innovation.spring.entry.memo.lv03.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record AuthSignInResponse(
+        String accessToken
+) {
+}
