+++ src/main/java/com/innovation/spring/entry/memo/lv03/exception/ExceptionResponse.java
@@ -1,0 +1,11 @@
+package com.innovation.spring.entry.memo.lv03.exception;
+
+import lombok.Builder;
+import org.springframework.http.HttpStatus;
+
+@Builder
+public record ExceptionResponse(
+        HttpStatus status,
+        String message
+) {
+}
