+++ src/main/java/com/innovation/spring/entry/memo/lv03/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.innovation.spring.entry.memo.lv03.repository;
+
+import com.innovation.spring.entry.memo.lv03.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    boolean existsByEmail(String email);
+
+    Optional<User> findByEmail(String email);
+
+}
