+++ src/main/java/com/innovation/spring/entry/memo/lv03/util/UriUtil.java
@@ -1,0 +1,10 @@
+package com.innovation.spring.entry.memo.lv03.util;
+
+import java.net.URI;
+
+public class UriUtil {
+
+    public static URI getSignInUri() {
+        return URI.create("api/auth/signin");
+    }
+}
